chore: upgrade to Python 3.10-13
As raised in #112, Python 3.9 support is already broken by some of
our type annotations and is about to go EOL anyway. Bumped our
specified requirements to 3.10+, added 3.13 to the CI builds.
`rich` also has a new major version 14 released, and I can't find any
changes that are breaking for our purpose, so broadened our dep range
on it to cover both v13 & v14.